backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pyjwt
requests
pydantic
passlib[bcrypt]
python-multipart
google-auth